requirements.txt
Django==1.8.4
django-cors-headers==1.1.0
MySQL-python==1.2.5
Naked==0.1.31
oauthlib==1.0.3
pycrypto==2.6.1
PyJWT==1.4.0
python-openid==2.2.5
python-social-auth==0.2.12
PyYAML==3.11
requests==2.7.0
requests-oauthlib==0.5.0
shellescape==3.4.1
six==1.9.0
uWSGI==2.0.11.1
wheel==0.24.0
